show feedback when user login in with invalid credentails on admin page

+++ GLF-FrontEnd-Admin-/src/screens/Login.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import "../styles/App.css";
+import { toast, Toaster } from "react-hot-toast";
 
 const AdminLoginScreen = () => {
   const [username, setUsername] = useState("");
@@ -13,27 +14,33 @@ const AdminLoginScreen = () => {
     try {
       setLoading(true);
 
-      // Send a POST request to your authentication endpoint
-      const response = await axios.post(`${process.env.REACT_APP_BACKEND_URL}/login`, {
-        username,
-        password,
-      });
+      const response = await axios.post(
+        `${process.env.REACT_APP_BACKEND_URL}/login`,
+        {
+          username,
+          password,
+        }
+      );
 
       // Assuming your server responds with a token upon successful login
       const token = response.data.token;
-
-      // Here you can handle the token, for example, store it in local storage
       localStorage.setItem("token", token);
       localStorage.setItem("username", username);
 
-      // Reset the form fields after login
       setUsername("");
       setPassword("");
       navigate("/");
     } catch (error) {
       console.error("Error logging in:", error);
-      localStorage.clear()
-      navigate("../login");
+      if (error.response && error.response.status !== 200) {
+        toast.error(error.response.data.message);
+        localStorage.clear();
+        navigate("/login");
+      } else {
+        toast.error("An error occurred. Please try again later.");
+      }
+    } finally {
+      setLoading(false);
     }
   };
 
@@ -46,6 +53,7 @@ const AdminLoginScreen = () => {
       ) : (
         <div className="flex items-center justify-center h-screen bg-gray-100">
           <div>
+            <Toaster position="top-right" reverseOrder={false} />
             <h1 className="text-2xl font-bold mb-4">Admin Login</h1>
 
             <div className="mb-4">
